requirements.txt
fastapi
requests
psutil
pyjwt
gunicorn
uvicorn[standard]
psycopg2-binary
